\name{na.itree}
\alias{na.itree}
\title{
  Handles Missing Values in an itree Object 
}
\usage{
na.itree(x)
}
\description{
  Handles missing values in an \code{itree} object.
  This function is identical to \code{na.rpart}. 
}
\arguments{
\item{x}{
  a model frame.
}}
\details{
  Internal function that handles missing values when calling the
  function \code{itree}.
}
\keyword{tree}
